requirements.txt
asgiref==3.9.1
bcrypt==4.1.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.1
redis==5.0.1
